Differentiation between discovery_port and discovery_local_port
When running IBRDTN in emulation mode, each IPND beacon is sent (using
UDP unicast) to a software hub, which then forwards this beacon to
each IBRDTN node individually. An IBRDTN node must therefore be able
to listen to a UDP port that may be different from the port number the
hub listens to.

An additional parameter 'discovery_local_port' is defined in the
configuration file, and a new function localPort() is defined in
Configuration.h and implemented in Configuration.cpp to return the
value of that parameter. In NativeDaemon.cpp function localPort() is
called to initialize variable disco_port, before creating an
IPNDAgent.

If parameter 'discovery_local_port' is not defined in the
configuration file, then function localPort() returns the same value
as function port(). This ensures compatibility with the standard
IBRDTN behavior, when emulation mode is not required.