Added support for dynamic port assignment in socket::tcpserversocket
In socket::tcpserversocket::up(), the socket is bound to an address
only if the port number in that address is != 0. Otherwise the port
number is assigned by the OS, and this port number is "discovered"
(using function get_port()) after we start listening to the
socket. This number is recorded in _address (which therefore cannot be
a declared as a 'const' anymore in socket.h).