Added support for dynamic port assignment in TCPConvergenceLayer
When several IBRDTN nodes must run concurrently in emulation mode on
the same host, with the TCP convergence layer enabled, they cannot all
listen to the same TCP port number. Dynamic port assigment is
therefore required for each node node to obtain its own port number to
listen to.

In TCPConvergenceLayer::add(...), support for dynamic port assignment
is based on the same method as that used in IPNDAgent. When a
tcpserversocket must be created to listen to port 0 (as specified in
the configuration file), then a dummy socket is first created with
dynamic port assignment. Then the port number assigned to this socket
is used as a reference to create all other sockets. The dummy socket
is then closed and deleted.