Added support for additional command-line parameters
When experiments involving hundreds of IBRDTN nodes are conducted
(especially in a scenario with churn), it is difficult to ensure the
repeatability of these experiments. New parameters have been
introduced in the daemon's configuration in order to provide better
control on the behavior of an IBRDTN node.

In subclass Configuration::Daemon, new parameters start_time,
stop_time, and seed are meant to determine respectively the time when
an IBRDTN daemon should start, the time when it should stop, and the
seed value to be used by its random generator. New options --start,
--stop, and --seed admitted in the command line allow the user to set
these parameters. There is no side-effect if these parameters are
missing.

In Main.cpp, a subclass TerminationThread has been defined that allows
the daemon to terminate spontaneously after a certain execution delay.

In function Main::main(), code has been introduced that sets the
random generator's seed (if option '--seed appears in the command
line), that defers the actual start of the daemon (if option '--start'
appears in the command line), and that starts TerminationThread (if
option '--stop' appears in the command line).